task(reducers): reducers/action factory for each stream
